Refactor: Split raster and vector routines.
